Refactor mappers and support querying/scanning index with paging
