+++ src/main/java/com/salesforce/dynamodbv2/mt/mappers/sharedtable/impl/TableMappingFactory.java
@@ -85,7 +85,7 @@ public class TableMappingFactory {
      * table is created, like the streamArn.
      */
     TableMapping getTableMapping(DynamoTableDescription virtualTableDescription) {
-        TableMapping tableMapping = new TableMapping(virtualTableDescription,
+        RandomPartitioningTableMapping tableMapping = new RandomPartitioningTableMapping(virtualTableDescription,
             createTableRequestFactory,
             secondaryIndexMapper,
             mtContext);
