+++ src/test/java/com/salesforce/dynamodbv2/ScanTest.java
@@ -1,15 +1,23 @@
 package com.salesforce.dynamodbv2;
 
 import static com.amazonaws.services.dynamodbv2.model.ComparisonOperator.EQ;
+import static com.amazonaws.services.dynamodbv2.model.ScalarAttributeType.N;
 import static com.amazonaws.services.dynamodbv2.model.ScalarAttributeType.S;
 import static com.salesforce.dynamodbv2.testsupport.ArgumentBuilder.MT_CONTEXT;
 import static com.salesforce.dynamodbv2.testsupport.ArgumentBuilder.ORGS_PER_TEST;
 import static com.salesforce.dynamodbv2.testsupport.DefaultTestSetup.TABLE1;
+import static com.salesforce.dynamodbv2.testsupport.DefaultTestSetup.TABLE3;
+import static com.salesforce.dynamodbv2.testsupport.ItemBuilder.GSI2_HK_FIELD;
+import static com.salesforce.dynamodbv2.testsupport.ItemBuilder.GSI2_RK_FIELD;
 import static com.salesforce.dynamodbv2.testsupport.ItemBuilder.HASH_KEY_FIELD;
 import static com.salesforce.dynamodbv2.testsupport.ItemBuilder.SOME_FIELD;
+import static com.salesforce.dynamodbv2.testsupport.TestSupport.GSI2_HK_FIELD_VALUE;
+import static com.salesforce.dynamodbv2.testsupport.TestSupport.GSI2_RK_FIELD_VALUE;
 import static com.salesforce.dynamodbv2.testsupport.TestSupport.HASH_KEY_OTHER_VALUE;
 import static com.salesforce.dynamodbv2.testsupport.TestSupport.HASH_KEY_VALUE;
+import static com.salesforce.dynamodbv2.testsupport.TestSupport.RANGE_KEY_OTHER_S_VALUE;
 import static com.salesforce.dynamodbv2.testsupport.TestSupport.SOME_FIELD_VALUE;
+import static com.salesforce.dynamodbv2.testsupport.TestSupport.SOME_OTHER_FIELD_VALUE;
 import static com.salesforce.dynamodbv2.testsupport.TestSupport.SOME_OTHER_OTHER_FIELD_VALUE;
 import static com.salesforce.dynamodbv2.testsupport.TestSupport.attributeValueToString;
 import static com.salesforce.dynamodbv2.testsupport.TestSupport.createAttributeValue;
@@ -43,6 +51,7 @@ import java.util.Map;
 import java.util.Random;
 import java.util.Set;
 import java.util.function.Function;
+import java.util.stream.Collectors;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ArgumentsSource;
 
@@ -167,6 +176,35 @@ class ScanTest {
         assertTrue(isFound, "all scans found no items... that's not right");
     }
 
+    @ParameterizedTest(name = "{arguments}")
+    @ArgumentsSource(DefaultArgumentProvider.class)
+    void scanGsiWithPaging(TestArgument testArgument) {
+        testArgument.forEachOrgContext(org -> {
+            // add another gsi2 record so there are multiple and we'd need paging
+            Map<String, AttributeValue> secondRecord = ItemBuilder
+                .builder(testArgument.getHashKeyAttrType(), HASH_KEY_VALUE)
+                .withDefaults()
+                .rangeKey(S, RANGE_KEY_OTHER_S_VALUE + "1")
+                .gsi2HkField(S, GSI2_HK_FIELD_VALUE)
+                .gsi2RkField(N, GSI2_RK_FIELD_VALUE + "1")
+                .build();
+            testArgument.getAmazonDynamoDb().putItem(new PutItemRequest().withTableName(TABLE3).withItem(secondRecord));
+
+            Map<String, AttributeValue> firstRecord = ItemBuilder
+                .builder(testArgument.getHashKeyAttrType(), HASH_KEY_VALUE)
+                .withDefaults()
+                .someField(S, SOME_OTHER_FIELD_VALUE + TABLE3 + org)
+                .build();
+            List<Map<String, AttributeValue>> expectedRecords = ImmutableList.of(firstRecord, secondRecord);
+
+            assertEquals(expectedRecords, executeScan(
+                exclusiveStartKey -> testArgument.getAmazonDynamoDb().scan(
+                    new ScanRequest(TABLE3)
+                        .withIndexName("testGsi2")
+                        .withLimit(1)
+                        .withExclusiveStartKey(exclusiveStartKey))));
+        });
+    }
 
     @ParameterizedTest(name = "{arguments}")
     @ArgumentsSource(ScanTestArgumentProvider.class)
