+++ src/test/java/com/salesforce/dynamodbv2/mt/mappers/sharedtable/impl/RandomPartitioningRecordMapperTest.java
@@ -17,12 +17,13 @@ import com.salesforce.dynamodbv2.mt.context.MtAmazonDynamoDbContextProvider;
 import com.salesforce.dynamodbv2.mt.mappers.CreateTableRequestBuilder;
 import com.salesforce.dynamodbv2.mt.mappers.index.DynamoSecondaryIndexMapperByTypeImpl;
 import com.salesforce.dynamodbv2.mt.mappers.metadata.DynamoTableDescriptionImpl;
+import java.util.Collections;
 import java.util.Optional;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
-class RecordMapperTest {
+class RandomPartitioningRecordMapperTest {
 
     private final String context = "context";
     private final String tableName = "table";
@@ -31,9 +32,9 @@ class RecordMapperTest {
     private final String virtualRk = "vrk";
     private final String physicalHk = "phk";
     private final String physicalRk = "prk";
-    private final ItemMapper itemMapper = new ItemMapper(
+    private final RandomPartitioningItemMapper itemMapper = new RandomPartitioningItemMapper(
         new StringFieldMapper(provider, tableName),
-        new TableMapping(new DynamoTableDescriptionImpl(
+        new RandomPartitioningTableMapping(new DynamoTableDescriptionImpl(
             CreateTableRequestBuilder.builder()
                 .withTableKeySchema(virtualHk, S, virtualRk, S).build()),
             new SingletonCreateTableRequestFactory(new DynamoTableDescriptionImpl(
@@ -42,9 +43,10 @@ class RecordMapperTest {
                 .getCreateTableRequest()),
             new DynamoSecondaryIndexMapperByTypeImpl(),
             null
-        ).getAllVirtualToPhysicalFieldMappings());
+        ).getTablePrimaryKeyFieldMappings(),
+        Collections.emptyMap());
     private final FieldMapper fieldMapper = new StringFieldMapper(provider, tableName);
-    private final RecordMapper sut = new RecordMapper(provider, tableName, itemMapper, fieldMapper, physicalHk);
+    private final RandomPartitioningRecordMapper sut = new RandomPartitioningRecordMapper(provider, tableName, itemMapper, fieldMapper, physicalHk);
 
     @ParameterizedTest
     @ValueSource(strings = { "context/table/abc", "context/table/0" })
